Update to use setup.py
Using setup.py to build the plugin makes it cleaner and simpler to
follow. No more bash tricks to find versions of things.

The installation directory for the plugin varies between Ubuntu and
CentOS so we're counting on glob for that now:

- Ubuntu: /usr/lib/pppd/$version/
- CentOS: /usr/lib64/pppd/$version/

This Makefile sets the PLUGINDIR in such a way that it works on both:

- PLUGINDIR: /usr/lib*/pppd/*/

The `chcon` commands have been commented out because I couldn't make
them work properly. Perhaps someome with better experience on SELinux
will be able to fix it later.